+++ services/serviceaccount/model_create_service_account_key_response_credentials.go
@@ -153,6 +153,33 @@ type CreateServiceAccountKeyResponseCredentialsGetSubArgType = string
 // Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
 type CreateServiceAccountKeyResponseCredentialsGetSubRetType = string
 
+/*
+	types and functions for tokenEndpoint
+*/
+
+// isNotNullableString
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+type CreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeType = *string
+
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func getCreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeTypeOk(arg CreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeType) (ret CreateServiceAccountKeyResponseCredentialsGetTokenEndpointRetType, ok bool) {
+	if arg == nil {
+		return ret, false
+	}
+	return *arg, true
+}
+
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func setCreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeType(arg *CreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeType, val CreateServiceAccountKeyResponseCredentialsGetTokenEndpointRetType) {
+	*arg = &val
+}
+
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+type CreateServiceAccountKeyResponseCredentialsGetTokenEndpointArgType = string
+
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+type CreateServiceAccountKeyResponseCredentialsGetTokenEndpointRetType = string
+
 // CreateServiceAccountKeyResponseCredentials struct for CreateServiceAccountKeyResponseCredentials
 // Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
 type CreateServiceAccountKeyResponseCredentials struct {
@@ -170,6 +197,8 @@ type CreateServiceAccountKeyResponseCredentials struct {
 	// Service account id
 	// REQUIRED
 	Sub CreateServiceAccountKeyResponseCredentialsGetSubAttributeType `json:"sub" required:"true"`
+	// Endpoint used to obtain access tokens for the service account.
+	TokenEndpoint CreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeType `json:"tokenEndpoint,omitempty"`
 }
 
 // Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
@@ -305,6 +334,33 @@ func (o *CreateServiceAccountKeyResponseCredentials) SetSub(v CreateServiceAccou
 	setCreateServiceAccountKeyResponseCredentialsGetSubAttributeType(&o.Sub, v)
 }
 
+// GetTokenEndpoint returns the TokenEndpoint field value if set, zero value otherwise.
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func (o *CreateServiceAccountKeyResponseCredentials) GetTokenEndpoint() (res CreateServiceAccountKeyResponseCredentialsGetTokenEndpointRetType) {
+	res, _ = o.GetTokenEndpointOk()
+	return
+}
+
+// GetTokenEndpointOk returns a tuple with the TokenEndpoint field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func (o *CreateServiceAccountKeyResponseCredentials) GetTokenEndpointOk() (ret CreateServiceAccountKeyResponseCredentialsGetTokenEndpointRetType, ok bool) {
+	return getCreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeTypeOk(o.TokenEndpoint)
+}
+
+// HasTokenEndpoint returns a boolean if a field has been set.
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func (o *CreateServiceAccountKeyResponseCredentials) HasTokenEndpoint() bool {
+	_, ok := o.GetTokenEndpointOk()
+	return ok
+}
+
+// SetTokenEndpoint gets a reference to the given string and assigns it to the TokenEndpoint field.
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func (o *CreateServiceAccountKeyResponseCredentials) SetTokenEndpoint(v CreateServiceAccountKeyResponseCredentialsGetTokenEndpointRetType) {
+	setCreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeType(&o.TokenEndpoint, v)
+}
+
 // Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
 func (o CreateServiceAccountKeyResponseCredentials) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
@@ -322,6 +378,9 @@ func (o CreateServiceAccountKeyResponseCredentials) ToMap() (map[string]interfac
 	}
 	if val, ok := getCreateServiceAccountKeyResponseCredentialsGetSubAttributeTypeOk(o.Sub); ok {
 		toSerialize["Sub"] = val
+	}
+	if val, ok := getCreateServiceAccountKeyResponseCredentialsGetTokenEndpointAttributeTypeOk(o.TokenEndpoint); ok {
+		toSerialize["TokenEndpoint"] = val
 	}
 	return toSerialize, nil
 }